+++ CHANGELOG.rst
@@ -16,10 +16,13 @@ Unreleased
 
 *
 
-0.1.0 – 2023-10-04
+0.1.0 – 2023-10-23
 **********************************************
 
+- Add Instructor Dashboard integration.
+- Add Superset XBlock.
+
 Added
-=====
+-----
 
 * First release on PyPI.